build-baremetal-toolchain.sh: Add --disable-libstdcxx-verbose to gcc2-nano build.
The --disable-libstdcxx-verbose prevents the gcc libraries from trying
to write debug garbage to stdout when things like terminate() are called
(say, if you throw an exception which is never caught).  This bloats the
firmware image and requires you to implement stub functions for things
like read(), write() and seek() - stuff that makes no sense in a small
firmware image.

Previously, in the gcc 9 series of ARM builds, this flag was set, but it
seems that the gnu-devtools-for-arm repo doesn't set them for some
reason - possibly a regression.