build-baremetal-toolchain.sh: Add support for gcc2-nano_eh and newlib-nano_eh builds.
These are the same as the nano builds except that exception handling is
enabled.  This is important because if exception handling is disabled
and you attempt to do a throw() the program will immediately terminate
because there is no exception information for how to propagate the
exception through the bottom-level throw() function!